// api/src/constants/routes.ts
// eslint-disable-next-line import/prefer-default-export
export enum ROUTES {
  admin = '/admin',
  signin = '/auth/login',
  signup = '/auth/register',
  logout = '/auth/logout',
  resetPassword = '/auth/reset-password',
  resetPasswordUpdate = '/auth/reset-password/update',
  logs = '/logs',
  tokenVerify = '/auth/verify',
  producer = '/producer',
  fazenda = '/fazenda',
  county = '/county',
  country = '/country',
  province = '/province',
  machine = '/machine',
  alfaia = '/alfaia',
  culture = '/culture',
  equipamento = '/equipamento',
  humanResouce = '/human-resource',
  infrastructure = '/infrastructure',
  meioEstacionario = '/meio-estacionario',
  cultureType = '/culture-types',
  equipamentoType = '/equipamento-types',
  humanResouceType = '/human-resource-types',
  infrastructureType = '/infrastructure-types',
  machineType = '/machine-types',
  meioEstacionarioType = '/meio-estacionario-types',
  animalType = '/animal-types',
  animal = '/animal',
  cooperative = '/cooperative',
  association = '/association',
  user = '/users',
  technician = '/technician',
  partner = '/partner',
  nationalProduction = '/national-production',
  consumption = '/consumption',
  marketPrices = '/market-prices',
}